use f64 based timeout instead of u64
So sub-second timeouts and other floating point values can be used.

+++ crates/cli/src/options.rs
@@ -195,7 +195,7 @@ struct BiteOpts {
     insecure: bool,
     /// request timeout in seconds
     #[arg(short, long, value_name = "SECONDS", default_value = "30")]
-    timeout: u64,
+    timeout: f64,
 }
 
 #[derive(Debug, Args)]
